Update README.md
DeLotGame is an ethereum-based prediction game where players pledge ETH to get DLT tokens to participate in guessing. The first time 1ETH = 1DLT, and the subsequent distribution is proportional. Players send ETH transactions containing hexadecimal data to participate in the game, the system will check whether the last bet matches the last character of the target block hash, and if it does, the reward will be calculated according to the formula (e.g., 1ETH + 4 bytes of data = 3.6ETH). ETH can be redeemed at any time, and the balance of the pool is allocated according to the percentage of coin holdings. The contract is protected by anti-re-entry protection, and the result is guaranteed to be fair by the block hash.